add earth event mapper

+++ src/main/java/com/joyful/earthquakes/model/EarthEvent.java
@@ -3,12 +3,13 @@ package com.joyful.earthquakes.model;
 import lombok.Builder;
 import lombok.Value;
 
+import java.time.LocalDateTime;
 import java.time.ZonedDateTime;
 
 @Value
 @Builder
 public class EarthEvent {
-    Integer magnitude;
+    Float magnitude;
     String country;
     String place;
     ZonedDateTime time;
